+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - `cluster list` command; Limes has removed multi-cluster support.
+- `--cluster` flag for domain and project subcommands.
 
 ## 2.0.1 - 2021-10-06